slice 테스트 파일 제거 후 reducer 테스트로 분리
