+++ src/App.test.js
@@ -1,10 +1,24 @@
 import { render, fireEvent } from '@testing-library/react';
 
-import '@testing-library/jest-dom';
+import { useDispatch, useSelector } from 'react-redux';
 
 import App from './App';
 
 describe('App', () => {
+  const dispatch = jest.fn();
+
+  beforeEach(() => {
+    dispatch.mockClear();
+    useDispatch.mockImplementation(() => dispatch);
+
+    useSelector.mockImplementation((selector) => selector({
+      player: {
+        userName: '',
+      },
+      summoner: {},
+    }));
+  });
+
   function renderApp() {
     return render(<App />);
   }
