+++ src/setupProxy.js
@@ -1,10 +1,10 @@
 const { createProxyMiddleware } = require('http-proxy-middleware');
 
-module.exports = function(app){
+module.exports = (app) => {
   app.use(
     createProxyMiddleware('/lol', {
       target: 'https://kr.api.riotgames.com',
-      changeOrigin: true
-    })
-  )
+      changeOrigin: true,
+    }),
+  );
 };
